% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minify_dir.R
\name{minifyJSDirectory}
\alias{minifyJSDirectory}
\alias{minifyCSSDirectory}
\alias{minifyHTMLDirectory}
\title{Minify File}
\usage{
minifyJSDirectory(
  directory,
  single_file = TRUE,
  minified_dir_name = directory,
  minified_file_name = paste0(basename(directory), ".min.js")
)

minifyCSSDirectory(
  directory,
  single_file = TRUE,
  minified_dir_name = directory,
  minified_file_name = paste0(basename(directory), ".min.css")
)

minifyHTMLDirectory(
  directory,
  single_file = TRUE,
  minified_dir_name = directory,
  minified_file_name = paste0(basename(directory), ".min.html")
)
}
\arguments{
\item{directory}{Path to the files to be minified}

\item{single_file}{Should all the files be minified to a single file? Default set to \code{TRUE}}

\item{minified_dir_name, minified_file_name}{If \code{single_file = TRUE}, the directory and name of the minified file}
}
\description{
Using the
}
\seealso{
\code{\link{minify}}
}
