% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minify.R
\name{minifyJSFile}
\alias{minifyJSFile}
\alias{minifyCSSFile}
\alias{minifyHTMLFile}
\title{Minify File}
\usage{
minifyJSFile(file_name)

minifyCSSFile(file_name)

minifyHTMLFile(file_name)
}
\arguments{
\item{file_name}{Path to the file to be minified}
}
\description{
Minify File
}
