% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minify.R
\name{minify}
\alias{minify}
\title{Minify Code}
\usage{
minify(code, language = MINIFY_LANGUAGES)
}
\arguments{
\item{code}{Snippet of code to minify}

\item{language}{Language of the code to minify. Options are \code{javascript}, \code{css}, \code{html}}
}
\description{
This...
}
\seealso{
\url{https://www.toptal.com/developers/javascript-minifier/api}
\url{https://www.toptal.com/developers/cssminifier/api}
\url{https://www.toptal.com/developers/html-minifier/api}
}
